rename package to follow java project group naming conventions
