+++ src/main/java/ai/rev/StreamContentType.java
@@ -1,4 +1,4 @@
-package revai;
+package ai.rev;
 
 import java.util.ArrayList;
 import java.util.List;
