+++ src/main/java/ai/rev/models/CustomVocabulary.java
@@ -1,4 +1,4 @@
-package revai.models;
+package ai.rev.models;
 
 import java.util.List;
