+++ src/main/java/ai/rev/models/streaming/StreamingResponseMessage.java
@@ -1,4 +1,4 @@
-package revai.models.streaming;
+package ai.rev.models.streaming;
 
 import com.google.gson.annotations.SerializedName;
